requirements.txt
paho-mqtt
matplotlib
pyyaml
PyQt6
